แก้ไข css และแจ้งเตือนตอนลบรายการในตะกร้า

+++ components/Navbar.vue
@@ -211,15 +211,12 @@ export default {
                 this.count_shop = 0;
                 this.Sum = 0;
                 localStorage.setItem('shoppingCart', JSON.stringify(newValue));
-                console.log("newValue", newValue);
                 this.dataValue = newValue;
                 this.dataValue.forEach((element, index) => {
                     this.Sum += element.product_price * element.amount;
                     this.count_shop = index + 1;
-                    console.log("index", index);
                 });
                 this.refresh()
-                console.log("2", this.dataValue);
             },
         },
     },
@@ -228,9 +225,7 @@ export default {
         this.dataValue.forEach((element, index) => {
             this.Sum += element.product_price * element.amount;
             this.count_shop = index + 1;
-            console.log("index", index);
         });
-        console.log("this.dataValue", this.dataValue);
     },
     methods: {
         async logout() {
@@ -270,25 +265,27 @@ export default {
                 cancelButtonText: 'ยกเลิก'
             }).then((result) => {
                 if (result.value == true) {
-                    this.$swal.fire(
-                        'success',
-                        'ลบสำเร็จ',
-                        'success'
-                    )
-                    const shoppingCart = this.dataValue;
-                    const productIndex = shoppingCart.findIndex(item => item.product_code === product.product_code);
-                    shoppingCart.splice(productIndex, 1);
-                    this.dataValue = shoppingCart;
-                    this.Sum = 0;
-                    this.dataValue.forEach(element => {
-                        this.Sum += element.product_price * element.amount;
-                    });
-                    // console.log("this.dataValue", this.dataValue);
-                    localStorage.setItem('shoppingCart', JSON.stringify(shoppingCart));
-                    this.$emit('update:modelValue', shoppingCart)
-                    setTimeout(() => {
-                        window.location.reload(true);
-                    }, 1000);
+                    this.$swal.fire({
+                        title: 'success',
+                        text: "ลบสำเร็จ",
+                        type: 'success',
+                        showConfirmButton: false,
+                        timer: 1500,
+                    }).then((e) => {
+                        if (e.dismiss == 'timer') {
+                            const shoppingCart = this.dataValue;
+                            const productIndex = shoppingCart.findIndex(item => item.product_code === product.product_code);
+                            shoppingCart.splice(productIndex, 1);
+                            this.dataValue = shoppingCart;
+                            this.Sum = 0;
+                            this.dataValue.forEach(element => {
+                                this.Sum += element.product_price * element.amount;
+                            });
+                            localStorage.setItem('shoppingCart', JSON.stringify(shoppingCart));
+                            this.$emit('update:modelValue', shoppingCart)
+                            window.location.reload(true);
+                        }
+                    })
                 }
             })
         }
